+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 ## master
 
 
+## 1. (..., 2017)
+
+* Remove simple server.
+* Fix browsersync.
+
 ## 1.0.1 (April 8, 2017)
 
 * Fix all typescript errors.